tools: updated VariantStatsCalculator and its tests.
