+++ biodata-models/src/main/java/org/opencb/biodata/models/variant/stats/VariantStats.java
@@ -275,7 +275,7 @@ public class VariantStats {
         return mendelianErrors;
     }
 
-    void setMendelianErrors(int mendelianErrors) {
+    public void setMendelianErrors(int mendelianErrors) {
         this.mendelianErrors = mendelianErrors;
     }
 
@@ -283,7 +283,7 @@ public class VariantStats {
         return casesPercentDominant;
     }
 
-    void setCasesPercentDominant(float casesPercentDominant) {
+    public void setCasesPercentDominant(float casesPercentDominant) {
         this.casesPercentDominant = casesPercentDominant;
     }
 
@@ -291,7 +291,7 @@ public class VariantStats {
         return controlsPercentDominant;
     }
 
-    void setControlsPercentDominant(float controlsPercentDominant) {
+    public void setControlsPercentDominant(float controlsPercentDominant) {
         this.controlsPercentDominant = controlsPercentDominant;
     }
 
@@ -299,7 +299,7 @@ public class VariantStats {
         return casesPercentRecessive;
     }
 
-    void setCasesPercentRecessive(float casesPercentRecessive) {
+    public void setCasesPercentRecessive(float casesPercentRecessive) {
         this.casesPercentRecessive = casesPercentRecessive;
     }
 
@@ -307,7 +307,7 @@ public class VariantStats {
         return controlsPercentRecessive;
     }
 
-    void setControlsPercentRecessive(float controlsPercentRecessive) {
+    public void setControlsPercentRecessive(float controlsPercentRecessive) {
         this.controlsPercentRecessive = controlsPercentRecessive;
     }
 
@@ -349,7 +349,7 @@ public class VariantStats {
         return passedFilters;
     }
 
-    void setPassedFilters(boolean passedFilters) {
+    public void setPassedFilters(boolean passedFilters) {
         this.passedFilters = passedFilters;
     }
 
@@ -357,7 +357,7 @@ public class VariantStats {
         return quality;
     }
 
-    void setQuality(float quality) {
+    public void setQuality(float quality) {
         this.quality = quality;
     }
 
@@ -365,7 +365,7 @@ public class VariantStats {
         return numSamples;
     }
 
-    void setNumSamples(int numSamples) {
+    public void setNumSamples(int numSamples) {
         this.numSamples = numSamples;
     }
 
@@ -386,234 +386,6 @@ public class VariantStats {
                 + ", casesPercentRecessive=" + casesPercentRecessive
                 + ", controlsPercentRecessive=" + controlsPercentRecessive
                 + '}';
-    }
-
-    public VariantStats calculate(Map<String, Map<String, String>> samplesData, Map<String, String> attributes, Pedigree pedigree) {
-        int[] allelesCount = new int[2];
-        int totalAllelesCount = 0, totalGenotypesCount = 0;
-
-        float controlsDominant = 0, casesDominant = 0;
-        float controlsRecessive = 0, casesRecessive = 0;
-
-        this.setNumSamples(samplesData.size());
-        this.setMissingAlleles(0);
-        this.setMissingGenotypes(0);
-        if (pedigree != null) {
-            this.setMendelianErrors(0);
-        }
-
-        for (Map.Entry<String, Map<String, String>> sample : samplesData.entrySet()) {
-            String sampleName = sample.getKey();
-            Genotype g = new Genotype(sample.getValue().get("GT"), this.getRefAllele(), this.getAltAllele());
-            this.addGenotype(g);
-
-            // Check missing alleles and genotypes
-            switch (g.getCode()) {
-                case ALLELES_OK:
-                    // Both alleles set
-                    allelesCount[g.getAllele(0)]++;
-                    allelesCount[g.getAllele(1)]++;
-
-                    totalAllelesCount += 2;
-                    totalGenotypesCount++;
-
-                    // Counting genotypes for Hardy-Weinberg (all phenotypes)
-                    if (g.isAlleleRef(0) && g.isAlleleRef(1)) { // 0|0
-                        this.getHw().incN_AA();
-                    } else if ((g.isAlleleRef(0) && g.getAllele(1) == 1) || (g.getAllele(0) == 1 && g.isAlleleRef(1))) {  // 0|1, 1|0
-                        this.getHw().incN_Aa();
-
-                    } else if (g.getAllele(0) == 1 && g.getAllele(1) == 1) {
-                        this.getHw().incN_aa();
-                    }
-
-                    break;
-                case HAPLOID:
-                    // Haploid (chromosome X/Y)
-                    allelesCount[g.getAllele(0)]++;
-                    totalAllelesCount++;
-                    break;
-                case MULTIPLE_ALTERNATES:
-                    // Alternate with different "index" than the one that is being handled
-                    break;
-                default:
-                    // Missing genotype (one or both alleles missing)
-                    this.setMissingGenotypes(this.getMissingGenotypes() + 1);
-                    if (g.getAllele(0) < 0) {
-                        this.setMissingAlleles(this.getMissingAlleles() + 1);
-                    } else {
-                        allelesCount[g.getAllele(0)]++;
-                        totalAllelesCount++;
-                    }
-
-                    if (g.getAllele(1) < 0) {
-                        this.setMissingAlleles(this.getMissingAlleles() + 1);
-                    } else {
-                        allelesCount[g.getAllele(1)]++;
-                        totalAllelesCount++;
-                    }
-                    break;
-
-            }
-
-            // Include statistics that depend on pedigree information
-            if (pedigree != null) {
-                if (g.getCode() == AllelesCode.ALLELES_OK || g.getCode() == AllelesCode.HAPLOID) {
-                    Individual ind = pedigree.getIndividual(sampleName);
-//                    if (MendelChecker.isMendelianError(ind, g, variant.getChromosome(), file.getSamplesData())) {
-//                        this.setMendelianErrors(this.getMendelianErrors() + 1);
-//                    }
-                    if (g.getCode() == AllelesCode.ALLELES_OK) {
-                        // Check inheritance models
-                        if (ind.getCondition() == Condition.UNAFFECTED) {
-                            if (g.isAlleleRef(0) && g.isAlleleRef(1)) { // 0|0
-                                controlsDominant++;
-                                controlsRecessive++;
-
-                            } else if ((g.isAlleleRef(0) && !g.isAlleleRef(1)) || (!g.isAlleleRef(0) || g.isAlleleRef(1))) { // 0|1 or 1|0
-                                controlsRecessive++;
-
-                            }
-                        } else if (ind.getCondition() == Condition.AFFECTED) {
-                            if (!g.isAlleleRef(0) && !g.isAlleleRef(1) && g.getAllele(0) == g.getAllele(1)) {// 1|1, 2|2, and so on
-                                casesRecessive++;
-                                casesDominant++;
-                            } else if (!g.isAlleleRef(0) || !g.isAlleleRef(1)) { // 0|1, 1|0, 1|2, 2|1, 1|3, and so on
-                                casesDominant++;
-
-                            }
-                        }
-
-                    }
-                }
-            } 
-
-        }  // Finish all samples loop
-        
-        // Set counts for each allele
-        this.setRefAlleleCount(allelesCount[0]);
-        this.setAltAlleleCount(allelesCount[1]);
-
-        // Calculate MAF and MGF
-        this.calculateAlleleFrequencies(totalAllelesCount);
-        this.calculateGenotypeFrequencies(totalGenotypesCount);
-
-        // Calculate Hardy-Weinberg statistic
-        this.getHw().calculate();
-
-        // Update variables finally used to update file_stats_t structure
-        if ("PASS".equalsIgnoreCase(attributes.get("FILTER"))) {
-            this.setPassedFilters(true);
-        }
-
-        if (attributes.containsKey("QUAL") && !(".").equals(attributes.get("QUAL"))) {
-            float qualAux = Float.valueOf(attributes.get("QUAL"));
-            if (qualAux >= 0) {
-                this.setQuality(qualAux);
-            }
-        }
-
-        if (pedigree != null) {
-            // Once all samples have been traversed, calculate % that follow inheritance model
-            controlsDominant = controlsDominant * 100 / (this.getNumSamples() - this.getMissingGenotypes());
-            casesDominant = casesDominant * 100 / (this.getNumSamples() - this.getMissingGenotypes());
-            controlsRecessive = controlsRecessive * 100 / (this.getNumSamples() - this.getMissingGenotypes());
-            casesRecessive = casesRecessive * 100 / (this.getNumSamples() - this.getMissingGenotypes());
-
-            this.setCasesPercentDominant(casesDominant);
-            this.setControlsPercentDominant(controlsDominant);
-            this.setCasesPercentRecessive(casesRecessive);
-            this.setControlsPercentRecessive(controlsRecessive);
-        }
-
-        return this;
-    }
-
-    /**
-     * Calculates the statistics for some variants read from a set of files, and 
-     * optionally given pedigree information. Some statistics like inheritance 
-     * patterns can only be calculated if pedigree information is provided.
-     * 
-     * @param variants The variants whose statistics will be calculated
-     * @param ped Optional pedigree information to calculate some statistics
-     */
-    public static void calculateStatsForVariantsList(List<Variant> variants, Pedigree ped) {
-        for (Variant variant : variants) {
-            for (VariantSourceEntry file : variant.getSourceEntries().values()) {
-                VariantStats stats = new VariantStats(variant).calculate(file.getSamplesData(), file.getAttributes(), ped);
-                file.setStats(stats); // TODO Correct?
-            }
-        }
-    }
-
-    private void calculateAlleleFrequencies(int totalAllelesCount) {
-        if (totalAllelesCount < 0) {
-            throw new IllegalArgumentException("The number of alleles must be equals or greater than zero");
-        }
-        
-        if (totalAllelesCount == 0) {
-            // Nothing to calculate here
-            this.setMaf(-1);
-            this.setMafAllele(null);
-            return;
-        }
-        
-        refAlleleFreq = refAlleleCount / (float) totalAllelesCount;
-        altAlleleFreq = altAlleleCount / (float) totalAllelesCount;
-        
-        if (refAlleleFreq <= altAlleleFreq) {
-            this.setMaf(refAlleleFreq);
-            this.setMafAllele(refAllele);
-        } else {
-            this.setMaf(altAlleleFreq);
-            this.setMafAllele(altAllele);
-        }
-    }
-
-    private void calculateGenotypeFrequencies(int totalGenotypesCount) {
-        if (totalGenotypesCount < 0) {
-            throw new IllegalArgumentException("The number of genotypes must be equals or greater than zero");
-        }
-        
-        if (genotypesCount.isEmpty() || totalGenotypesCount == 0) {
-            // Nothing to calculate here
-            this.setMgf(-1);
-            this.setMgfGenotype(null);
-            return;
-        }
-        
-        // Set all combinations of genotypes to zero
-        genotypesFreq.put(new Genotype("0/0", refAllele, altAllele), 0.0f);
-        genotypesFreq.put(new Genotype("0/1", refAllele, altAllele), 0.0f);
-        genotypesFreq.put(new Genotype("1/1", refAllele, altAllele), 0.0f);
-        
-        // Insert the genotypes found in the file
-        for (Map.Entry<Genotype, Integer> gtCount : genotypesCount.entrySet()) {
-            if (gtCount.getKey().getCode() == AllelesCode.ALLELES_MISSING) {
-                // Missing genotypes shouldn't have frequencies calculated
-                continue;
-            }
-            
-            float freq = (totalGenotypesCount > 0) ? gtCount.getValue() / (float) totalGenotypesCount : 0;
-            genotypesFreq.put(gtCount.getKey(), freq);
-        }
-        
-        // Traverse the genotypes to see which one has the MGF
-        float currMgf = Float.MAX_VALUE;
-        Genotype currMgfGenotype = null;
-        
-        for (Map.Entry<Genotype, Float> gtCount : genotypesFreq.entrySet()) {
-            float freq = gtCount.getValue();
-            if (freq < currMgf) {
-                currMgf = freq;
-                currMgfGenotype = gtCount.getKey();
-            }
-        }
-        
-        if (currMgfGenotype != null) {
-            this.setMgf(currMgf);
-            this.setMgfGenotype(currMgfGenotype.toString());
-        }
     }
 
     @Override
